Move watchEpilogue and ChronicleView.toJson to WebHumanIO
